feat : add RateDiscountPolicy
할인정책 적용

Role이 User라면 10% 할인 및 다형성을 사용하여
재사용성을 높임.

+++ src/main/java/org/hanghae/markethub/domain/discount/DiscountPolicy.java
@@ -9,5 +9,5 @@ public interface DiscountPolicy {
     /*
      * @return 할인 대상 금액
      */
-    BigDecimal discount (User user, int price);
+    BigDecimal discount (User user, BigDecimal price);
 }
